+++ NEWS.md
@@ -7,6 +7,7 @@ Announcements:
  * Change work in progress jobs endpoint from `[..]/job/wip` to `[..]/jobs-wip`
  * Documentation updates for Docs repo issue #840, GPKG Export.
  * Fix SHP exports, now it uses "the_geom" column by default when a dataset has more than one geometry column.
+ * Logging all errors
 
 
 ## 1.47.1